Insurance choice code sort of works
Solver breaks after a couple periods with insurance choice, but the code
has no immediate bugs.  Need to test with low pLvls back in, but will
probably need to use a smarter method for the JDfix in order to save
memory.  Will move on to testing simulation, multithreading, and basic
calibration for now.